Init a library crate based on last year's
I've carried forward the input file parsing + CLI args logic. The arg parsing and file I/O basically has to be done by the host... and I don't see the value in placing the entire input file as-is into a buffer and using compute kernels to parse. Let's just say parsing the input files and laying out an in-memory representation of the problem is NOT the 'main' work