Solved Day 1 Part 1
Basically all of the computation is done in OpenCL :)

I also made the classic mistake of using a readwrite buffer with reads across group boundaries -> race conditions

I think running into that in a real algorithm though is a good learning opportunity. Did the simple fix of just having a second buffer with the relevant data for reading (smaller than the whole read/write buffer).